auth-temp/auth - Copy/requirements.txt
Flask
Flask-Login
Flask-Mail
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
python-dotenv
